+++ src/main/java/me/myogoo/ae2tangiblebookmarks/client/TranslateKey.java
@@ -1,9 +1,0 @@
-package me.myogoo.ae2tangiblebookmarks.client;
-
-public class TranslateKey {
-    public static String CATEGORY = "key.categories.ae2tb";
-
-    public static String PICKUP_SET_ITEM = "key.ae2tb.pickup_set_item";
-    public static String PICKUP_SINGLE_ITEM = "key.ae2tb.pickup_single_item";
-    public static String PICKED_ITEM_AUTOCRAFTING = "key.ae2tb.picked_item_autocrafting";
-}