+++ src/main/java/me/myogoo/ae2tb/config/AE2TBConfig.java
@@ -1,0 +1,10 @@
+package me.myogoo.ae2tb.config;
+
+import net.minecraftforge.common.ForgeConfigSpec;
+
+public final class AE2TBConfig {
+    public static final ForgeConfigSpec CLIENT_SPEC = new ForgeConfigSpec.Builder().build();
+
+    private AE2TBConfig() {
+    }
+}