+++ src/main/java/me/myogoo/ae2tb/init/AE2TBConfig.java
@@ -1,0 +1,14 @@
+package me.myogoo.ae2tb.init;
+
+import net.minecraftforge.fml.ModLoadingContext;
+import net.minecraftforge.fml.config.ModConfig;
+
+public final class AE2TBConfig {
+    private AE2TBConfig() {
+    }
+
+    public static void initialize() {
+        ModLoadingContext.get().registerConfig(ModConfig.Type.CLIENT, me.myogoo.ae2tb.config.AE2TBConfig.CLIENT_SPEC,
+                "ae2tb-client.toml");
+    }
+}