+++ CHANGELOG.md
@@ -1,2 +1,11 @@
-# Feature
-- port 1.20.1
+# Changelog
+## Feature
+- Now need Terminal Bookmark Interact Card for interact JEI/EMI/REI's bookamrk.
+- Add Terminal Config Tab
+- Add guideme
+
+- Add Translate zh_cn. thanks to @ChuijkYahus
+
+## Bug Fixes
+- Fixed an issue in REI where input would not register if the key binding did not include a mouse input
+- Fix Crash when server load